challenge dialog moved to containers
