+++ stem-explorer-ng/src/app/containers/challenge-dialog/challenge-dialog.component.spec.ts
